+++ tasks.py
@@ -13,41 +13,33 @@ def clean(c):
         "*.pyc",
     ]
     for pattern in patterns:
-        c.run("find . -d -name '{}' -exec rm -rf {{}} \\;".format(pattern), pty=True, echo=cmd_echo)
+        c.run("find . -d -name '{}' -exec rm -rf {{}} \\;".format(pattern), echo=cmd_echo)
 
     with c.cd("docs"):
-        c.run("poetry run -- make clean", pty=True, echo=cmd_echo)
+        c.run("make clean", echo=cmd_echo)
 
 
 @task
 def lint(c):
-    c.run("poetry run -- mypy basana", pty=True, echo=cmd_echo)
-    c.run("poetry run -- ruff check", pty=True, echo=cmd_echo)
+    c.run("mypy basana", echo=cmd_echo)
+    c.run("ruff check", echo=cmd_echo)
 
 
 @task(lint)
 def test(c, html_report=False):
     # Execute testcases.
-    cmd = "poetry run -- pytest -vv --cov --cov-config=setup.cfg --durations=10"
+    cmd = "pytest -vv --cov --cov-config=setup.cfg --durations=10"
     if html_report:
         cmd += " --cov-report=html:cov_html"
-    c.run(cmd, pty=True, echo=cmd_echo)
-
-
-@task
-def create_virtualenv(c, all_extras=True):
-    cmd = ["poetry", "install"]
-    if all_extras:
-        cmd.append("--all-extras")
-    c.run(" ".join(cmd), pty=True, echo=cmd_echo)
+    c.run(cmd, echo=cmd_echo)
 
 
 @task
 def build_docs(c):
     with c.cd("docs"):
-        c.run("poetry run -- make html", pty=True, echo=cmd_echo)
+        c.run("make html", echo=cmd_echo)
 
 
 @task
 def build_dist(c):
-    c.run("poetry build --clean", pty=True, echo=cmd_echo)
+    c.run("rm -rf dist && uv build", echo=cmd_echo)